% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{cranlogs_cum}
\alias{cranlogs_cum}
\title{Get the daily downloads of a package in the recent one year}
\usage{
cranlogs_cum(x)
}
\arguments{
\item{x}{Package name}
}
\value{
a data frame
}
\description{
Get the daily downloads of a package in the recent one year
}
\examples{
c('Epi', 'EpiModel', 'epibasix', 'epiDisplay', 'epiR', 'epitools', 'nCov2019', 'pubh', 'survival', 'survey', 'surveillance', 'SPARSEMODr', 'WHO') |>
  lapply(cranlogs_cum) |>
  bind_rows() |>
  ggplot() +
  geom_line(aes(date, CumDownload)) +
  theme(axis.text.x = element_text(angle = 90)) +
  scale_y_continuous(labels = function(x) format(x, scientific = TRUE)) +
  facet_wrap(~package, scales = 'free_y')
}
