% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{create_gannt_txt}
\alias{create_gannt_txt}
\title{Create text for a Gannt chart}
\usage{
create_gannt_txt(x)
}
\arguments{
\item{x}{}
}
\description{
Create text for a Gannt chart
}
\examples{
x <- tibble::tribble(~Start, ~End, ~Activity, ~Status, "2021-01-01", "2021-03-31",
    "Research design", "done", "2021-04-01", "2021-09-30", "Lab preparation",
    "done", "2021-10-01", "2021-10-31", "Outdoor test", "active", "2021-11-01",
    "2023-01-31", "Collect data", "crit, active", "2021-05-01", "2022-04-30",
    "R package dev.", "active", "2023-02-01", "2023-08-31", "Analyze data", "",
    "2023-06-01", "2024-01-01", "Manuscript", "")
library(DiagrammeR)
gannt_txt <- paste("gantt", "dateFormat  YYYY-MM-DD", "section Basic Tasks",
    create_gannt_txt(x[1:4, ]), "section Products", create_gannt_txt(x[5:7, ]),
    sep = "\n")
mermaid(gannt_txt)
}
