% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{gs_journal}
\alias{gs_journal}
\title{Retrieve the cites, impact factor (IF), the Eigenfactor score, the rank, the ISSN, the publisher, and so on, of a journal.}
\usage{
gs_journal(journal)
}
\arguments{
\item{journal}{}
}
\value{
a list
}
\description{
Retrieve the cites, impact factor (IF), the Eigenfactor score, the rank, the ISSN, the publisher, and so on, of a journal.
}
