% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{gs_person}
\alias{gs_person}
\title{Retrieve historical citation data, the network of co-authors, the total publication number, the total distinct journal number, the top journal number, the year of the oldest article, the profile information, and the publication list for a scientist.}
\usage{
gs_person(id)
}
\arguments{
\item{id}{Google Scholar ID}
}
\value{
a list
}
\description{
Retrieve historical citation data, the network of co-authors, the total publication number, the total distinct journal number, the top journal number, the year of the oldest article, the profile information, and the publication list for a scientist.
}
\examples{
gsid <- "JoN88RAAAAAJ"
ls_person <- gs_person(gsid)
ls_journal <- gs_journal(ls_person$pub$journal[1])
ls_pub <- gs_pub(gsid, ls_person$pub$pubid[1])
}
