% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{gs_pub}
\alias{gs_pub}
\title{Retrieve the historic annual citations and the complete author list of a publication.}
\usage{
gs_pub(id, pubid)
}
\arguments{
\item{pubid}{}
}
\description{
Retrieve the historic annual citations and the complete author list of a publication.
}
