% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{plot_gannt}
\alias{plot_gannt}
\title{Plot a Gannt chart with ggplot2}
\usage{
plot_gannt(x)
}
\arguments{
\item{x}{}
}
\description{
Plot a Gannt chart with ggplot2
}
\examples{
library(ggplot2)
x <- tibble::tribble(~Start, ~End, ~Activity, ~Status, "2021-01-01", "2021-03-31",
    "Research design", "done", "2021-04-01", "2021-09-30", "Lab preparation", "done",
    "2021-10-01", "2021-10-31", "Outdoor test", "active", "2021-11-01", "2023-01-31",
    "Collect data", "crit, active", "2021-05-01", "2022-04-30", "R package dev.",
    "active", "2023-02-01", "2023-08-31", "Analyze data", "", "2023-06-01", "2024-01-01",
    "Manuscript", "")
library(tidyr)
plot_gannt(x)
}
