% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph206fun.R
\name{roxygen_colname}
\alias{roxygen_colname}
\title{Generate roxygen items for a data frame}
\usage{
roxygen_colname(x)
}
\arguments{
\item{x}{a dataframe}
}
\description{
Generate roxygen items for a data frame
}
